fix: resolve JSX parsing issues and pass all tests
